Enhancing Inventory
Equipping objects now removes it from your inventory and puts it in its respective equipped slot. It will swap what is in your inventory with what is currently equipped unless you are equipping to an empty slot.

The check command now also display the same information as the equipping command. There are some tabbing errors, but they can be fixed later.